added guild_categories and database standards and converted all value types to reflect db (#2)

+++ internal/connections/communications.go
@@ -26,7 +26,7 @@ func sendPacketToAll(packet packets.BasePacket) {
 	waitGroup.Wait()
 }
 
-func sendSystemMessageViaConn(conn *websocket.Conn, severity types.Severity, message string, channelId int) {
+func sendSystemMessageViaConn(conn *websocket.Conn, severity types.Severity, message string, channelId uint64) {
 	sysMessage := packets.SystemMessage{
 		Severity:  severity,
 		Message:   message,
