+++ internal/connections/manager.go
@@ -8,21 +8,21 @@ import (
 )
 
 var (
-	Map   = make(map[int]*websocket.Conn)
+	Map   = make(map[uint64]*websocket.Conn)
 	MapMu sync.Mutex
 )
 
 // Note that it is assumed that a user is verified if added to the connections map.
 // Since a user ID is provided to add a connection, then checks have already been ran.
 
-func addNewConnection(userId int, conn *websocket.Conn) {
+func addNewConnection(userId uint64, conn *websocket.Conn) {
 	MapMu.Lock()
 	Map[userId] = conn
 	MapMu.Unlock()
 	log.Printf("User %d connected", userId)
 }
 
-func removeConnection(userId int) {
+func removeConnection(userId uint64) {
 	MapMu.Lock()
 	delete(Map, userId)
 	MapMu.Unlock()
